+++ src/test/java/racinggame/cars/CarNameTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import racinggame.common.Messages;
+
 class CarNameTest {
 
 	@Test
