+++ src/test/java/racinggame/cars/CarTest.java
@@ -3,13 +3,14 @@ package racinggame.cars;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
-import static racinggame.cars.Constant.*;
+import static racinggame.common.Constant.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
 
 import nextstep.utils.Randoms;
+import racinggame.common.Messages;
 
 public class CarTest {
