+++ src/test/java/racinggame/game/GameCountTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import racinggame.cars.Messages;
+import racinggame.common.Messages;
 
 class GameCountTest {
